:recycle: split assign button
